Restructure: reorganized packages, removed pkgType in DependencyNode.
