+++ src/main/java/com/jfrog/ide/common/components/IssueNode.java
@@ -1,4 +1,4 @@
-package com.jfrog.ide.common.tree;
+package com.jfrog.ide.common.components;
 
 import javax.swing.tree.TreeNode;
