+++ src/main/java/com/jfrog/ide/common/components/VulnerabilityNode.java
@@ -1,12 +1,15 @@
-package com.jfrog.ide.common.tree;
+package com.jfrog.ide.common.components;
 
+import com.jfrog.ide.common.components.subentities.Cve;
+import com.jfrog.ide.common.components.subentities.ResearchInfo;
+import com.jfrog.ide.common.components.subentities.Severity;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static com.jfrog.ide.common.tree.Severity.getNotApplicableSeverity;
+import static com.jfrog.ide.common.components.subentities.Severity.getNotApplicableSeverity;
 
 /**
  * @author yahavi
