+++ src/main/java/com/jfrog/ide/common/components/subentities/License.java
@@ -1,4 +1,4 @@
-package com.jfrog.ide.common.tree;
+package com.jfrog.ide.common.components.subentities;
 
 public class License {
     private final String name;
