+++ src/main/java/com/jfrog/ide/common/scan/GraphScanLogic.java
@@ -1,9 +1,14 @@
 package com.jfrog.ide.common.scan;
 
+import com.jfrog.ide.common.components.DependencyNode;
+import com.jfrog.ide.common.components.LicenseViolationNode;
+import com.jfrog.ide.common.components.VulnerabilityNode;
+import com.jfrog.ide.common.components.subentities.License;
+import com.jfrog.ide.common.components.subentities.ResearchInfo;
+import com.jfrog.ide.common.components.subentities.Severity;
+import com.jfrog.ide.common.components.subentities.SeverityReason;
 import com.jfrog.ide.common.configuration.ServerConfig;
 import com.jfrog.ide.common.log.ProgressIndicator;
-import com.jfrog.ide.common.tree.License;
-import com.jfrog.ide.common.tree.*;
 import com.jfrog.xray.client.Xray;
 import com.jfrog.xray.client.services.common.Cve;
 import com.jfrog.xray.client.services.scan.*;
@@ -35,11 +40,9 @@ import static org.apache.commons.lang3.StringUtils.*;
 @Setter
 public class GraphScanLogic implements ScanLogic {
     public static final String MINIMAL_XRAY_VERSION_SUPPORTED_FOR_GRAPH_SCAN = "3.29.0";
-    private String pkgType;
     private Log log;
 
-    public GraphScanLogic(String pkgType, Log log) {
-        this.pkgType = pkgType;
+    public GraphScanLogic(Log log) {
         this.log = log;
     }
 
@@ -211,7 +214,7 @@ public class GraphScanLogic implements ScanLogic {
 
     private void addVulnerabilityResult(Map<String, DependencyNode> results, Vulnerability vulnerability, String watchName) {
         for (Map.Entry<String, ? extends Component> entry : vulnerability.getComponents().entrySet()) {
-            DependencyNode dependencyNode = getDependency(results, entry);
+            DependencyNode dependencyNode = getDependency(results, entry.getKey());
 
             if (vulnerability.getCves() == null || vulnerability.getCves().size() == 0) {
                 VulnerabilityNode vulnerabilityNode = convertToIssue(vulnerability, entry.getValue(), null, watchName);
@@ -246,13 +249,13 @@ public class GraphScanLogic implements ScanLogic {
         return new VulnerabilityNode(vulnerability.getIssueId(), Severity.valueOf(vulnerability.getSeverity()),
                 StringUtils.defaultIfBlank(vulnerability.getSummary(), "N/A"), component.getFixedVersions(),
                 component.getInfectedVersions(),
-                new com.jfrog.ide.common.tree.Cve(cveId, cvssV2Score, cvssV2Vector, cvssV3Score, cvssV3Vector),
+                new com.jfrog.ide.common.components.subentities.Cve(cveId, cvssV2Score, cvssV2Vector, cvssV3Score, cvssV3Vector),
                 vulnerability.getEdited(), watchNames, vulnerability.getReferences(), researchInfo);
     }
 
     private void addLicenseViolationResult(Map<String, DependencyNode> results, Violation licenseViolation) {
         for (Map.Entry<String, ? extends Component> entry : licenseViolation.getComponents().entrySet()) {
-            DependencyNode dependencyNode = getDependency(results, entry);
+            DependencyNode dependencyNode = getDependency(results, entry.getKey());
             List<String> watchNames = null;
             if (licenseViolation.getWatchName() != null) {
                 watchNames = Collections.singletonList(licenseViolation.getWatchName());
@@ -264,9 +267,8 @@ public class GraphScanLogic implements ScanLogic {
         }
     }
 
-    private DependencyNode getDependency(Map<String, DependencyNode> results, Map.Entry<String, ? extends Component> compEntry) {
-        String componentId = compEntry.getKey();
-        results.putIfAbsent(componentId, new DependencyNode().componentId(componentId).pkgType(pkgType));
+    private DependencyNode getDependency(Map<String, DependencyNode> results, String componentId) {
+        results.putIfAbsent(componentId, new DependencyNode().componentId(componentId));
         return results.get(componentId);
     }
